Custom Made .gitignore, 1 class exercise, and READMEs
Mostly Setup,
custom Made .gitignore that is made to meet the expectations for the directory design that I use,
a-bit of what I had to do in class for my DNP and a-bit of setting up READMEs,
I still have to populate them with data.
the Custom Made .gitignore could be optimized further but for now Is stable and has data more than enough
as I made sure to include future possible files conflicts according to what is standard to have or what
I thought will possibly be meet by me during exercises from the IDEs or the plugins that I will use,
Is made to come back to it if I run in any conflicts during exercises.